item_valor_service/requirements.txt
Flask==2.2.2
Flask-JWT-Extended==4.4.4
requests==2.28.1
Flask-SQLAlchemy==3.0.2
Werkzeug==2.2.2
psycopg2-binary==2.9.3
pulsar-client==3.3.0